String class references to use ::class
Updated the class references that were written as string to use the class name resolution constant.
The feature was added in PHP 5.5 and helps IDEs recognising the references, making it possible to
jump to references, search for usages, etc.

I don't know if there is a reason it is not used currently, but I couldn't think of a reason so here is
a pull request :)

ref: http://php.net/manual/en/language.oop5.basic.php#language.oop5.basic.class.class